Batch

+++ src/main/resources/db/initDB.sql
@@ -5,7 +5,7 @@ CREATE TABLE  students (
   stu_name   VARCHAR(80),
   stu_major  VARCHAR(80),
   stu_type   VARCHAR(80),
-  usual_grade  INT,
-  design_grade INT,
-  exam_grade   INT
+  usual_grade  DOUBLE ,
+  design_grade DOUBLE,
+  exam_grade   DOUBLE
 )
